pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deceit_env"
version = "0.1.0"
description = "RL environment that trains LLMs to be honest under adversarial pressure"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "openenv-core[core]>=0.2.1",
    "pytest>=7.0",
    "python-dotenv",
    "openai>=1.0",
    "fastapi",
    "uvicorn",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
